+++ studio/src/app/services/deck/deck-dashboard.service.ts
@@ -1,9 +1,9 @@
 import firebase from 'firebase/app';
 import 'firebase/firestore';
 
-import {Deck, DeckData} from '../../../models/data/deck';
+import {Deck, DeckData} from '../../models/data/deck';
 
-import {DeckService} from '../../data/deck/deck.service';
+import {DeckService} from '../data/deck/deck.service';
 
 export interface DeckDashboardCloneResult {
   from: Deck;
